Consolidate lexer code & dedupe Kind predicates
* Consolidate all Kind-based predicates into kind.jl and deduplicate the
  functions from Tokenize which are duplicated in JuliaSyntax. Minor
  restyling of the names here for consistency with the other predicates
  in JuliaSyntax.jl.
* Move all of Tokenize into a pair of files and remove the sub-modules
  as these don't really seem to achieve much now that we use K_str and
  don't need a namespace to contain the token kinds.
